requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llama-cloud==0.1.26
llama-cloud-services==0.6.32
llama-index==0.12.42
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.42
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.32
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.6
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wrapt==1.17.2
yarl==1.20.1
